Moved get_working_dir and get_project_root to State w/ updated tests
